refactor: commonize formatString, SfppHasher, and CryptoCodec
Move platform-specific implementations to pure Kotlin in commonMain:

- formatString: Move pure-Kotlin format parser from iosMain to
  commonMain, eliminating expect/actual and jvmAndroid String.format()
  wrapper. All 60+ call sites now use a single implementation.

- SfppHasher: Replace ByteBuffer/ByteOrder (JVM) with pure Kotlin
  little-endian byte encoding. Use Okio ByteString.sha256() for
  cross-platform hashing. Eliminates expect object entirely.

- CryptoCodec: Replace MessageDigest (JVM) and CoreCrypto (iOS) with
  Okio ByteString.sha256(). ZlibCodec remains as expect/actual (no
  pure-Kotlin zlib available). Renamed CodecActual.kt → ZlibCodec.kt.

Net: -82 lines, 5 platform-specific files eliminated.
