+++ core/model/src/commonMain/kotlin/org/meshtastic/core/model/util/SfppHasher.kt
@@ -16,7 +16,27 @@
  */
 package org.meshtastic.core.model.util
 
+import okio.ByteString.Companion.toByteString
+
 /** Computes SFPP (Store-Forward-Plus-Plus) message hashes for deduplication. */
-expect object SfppHasher {
-    fun computeMessageHash(encryptedPayload: ByteArray, to: Int, from: Int, id: Int): ByteArray
+object SfppHasher {
+    private const val HASH_SIZE = 16
+    private const val INT_BYTES = 4
+    private const val INT_COUNT = 3
+    private const val SHIFT_8 = 8
+    private const val SHIFT_16 = 16
+    private const val SHIFT_24 = 24
+
+    fun computeMessageHash(encryptedPayload: ByteArray, to: Int, from: Int, id: Int): ByteArray {
+        val input = ByteArray(encryptedPayload.size + INT_BYTES * INT_COUNT)
+        encryptedPayload.copyInto(input)
+        var offset = encryptedPayload.size
+        for (value in intArrayOf(to, from, id)) {
+            input[offset++] = value.toByte()
+            input[offset++] = (value shr SHIFT_8).toByte()
+            input[offset++] = (value shr SHIFT_16).toByte()
+            input[offset++] = (value shr SHIFT_24).toByte()
+        }
+        return input.toByteString().sha256().toByteArray().copyOf(HASH_SIZE)
+    }
 }
